backend/requirements.txt
torcheval
captum
numpy
matplotlib
pandas
scikit-learn
opencv-python
pillow
python-dotenv
dotenv
flask
flask-cors
jupyter
notebook
tqdm
seaborn
pandas

